Group composition components into folders
