+++ HackrNewsApp/HackrNewsApp/Common/ContentType.swift
