+++ HackrNewsApp/HackrNewsApp/Common/MainQueueDispatchDecorator.swift
@@ -1,0 +1,20 @@
+//
+//  Copyright © 2021 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+
+final class MainQueueDispatchDecorator<T> {
+    let decoratee: T
+
+    init(_ decoratee: T) {
+        self.decoratee = decoratee
+    }
+
+    func dispatch(completion: @escaping () -> Void) {
+        guard Thread.isMainThread else {
+            return DispatchQueue.main.async(execute: completion)
+        }
+        completion()
+    }
+}
