+++ HackrNewsApp/HackrNewsApp/Common/WeakRefVirtualProxy.swift
@@ -1,0 +1,13 @@
+//
+//  Copyright © 2021 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+
+final class WeakRefVirtualProxy<T: AnyObject> {
+    weak var object: T?
+
+    init(_ object: T) {
+        self.object = object
+    }
+}
